+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Improvements
 - Started using SpriteBatchContext for Draw and DrawTransformed methods
 - Make use of the new consuming variants in InputHandler and Keybind to consume UiControls inputs
 - Allow Tooltip to manage more than one paragraph and make it easier to add new lines
+- Allow adding dropdown elements at a specified index
 
 Fixes
 - Fixed auto-nav tooltip displaying on the selected element even when not in auto-nav mode